+++ soup/wasm.hpp
@@ -120,8 +120,8 @@ namespace soup
 			size_t stack_size;
 		};
 
-		bool skipOverBranch(Reader& r, size_t depth = 0) SOUP_EXCAL;
-		[[nodiscard]] bool doBranch(Reader& r, size_t depth, std::stack<CtrlFlowEntry>& ctrlflow) SOUP_EXCAL;
+		bool skipOverBranch(Reader& r, uint32_t depth = 0) SOUP_EXCAL;
+		[[nodiscard]] bool doBranch(Reader& r, uint32_t depth, std::stack<CtrlFlowEntry>& ctrlflow) SOUP_EXCAL;
 		[[nodiscard]] bool doCall(size_t type_index, size_t function_index) SOUP_EXCAL;
 	};
 }
